requirements.txt
###### Requirements without Version Specifiers ######
numpy
grpcio
grpcio-tools
pygame
pyinstrument
pygame-textinput
win-precise-time
#

